# Source/Shared/Iwentys.Sdk/Connected Services/Iwentys.Sdk/OpenAPI.nswag.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Iwentys.Endpoint.Server",
    "version": "1.0"
  },
  "paths": {
    "/api/achievements/students/{studentId}": {
      "get": {
        "tags": [
          "Achievement"
        ],
        "operationId": "students",
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AchievementInfoDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AchievementInfoDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AchievementInfoDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/assignments": {
      "get": {
        "tags": [
          "Assignment"
        ],
        "operationId": "assignmentsAll",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssignmentInfoDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssignmentInfoDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssignmentInfoDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Assignment"
        ],
        "operationId": "assignments",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignmentCreateArguments"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignmentCreateArguments"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AssignmentCreateArguments"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentInfoDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentInfoDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentInfoDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/assignments/{assignmentId}/complete": {
      "get": {
        "tags": [
          "Assignment"
        ],
        "operationId": "complete",
        "parameters": [
          {
            "name": "assignmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Github/student/{studentId}": {
      "get": {
        "tags": [
          "Github"
        ],
        "operationId": "student",
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CodingActivityInfoResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CodingActivityInfoResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CodingActivityInfoResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/quests": {
      "post": {
        "tags": [
          "Quest"
        ],
        "operationId": "quests2",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuestRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuestRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuestRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/QuestInfoDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestInfoDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestInfoDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Student": {
      "put": {
        "tags": [
          "Student"
        ],
        "operationId": "Student",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentUpdateRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentUpdateRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StudentUpdateRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/IwentysUserInfoDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IwentysUserInfoDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/IwentysUserInfoDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tournaments/{id}": {
      "get": {
        "tags": [
          "Tournament"
        ],
        "operationId": "tournaments",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TournamentInfoResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TournamentInfoResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TournamentInfoResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AchievementInfoDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "gettingTimeUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "IwentysUserInfoDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "middleName": {
            "type": "string",
            "nullable": true
          },
          "secondName": {
            "type": "string",
            "nullable": true
          },
          "isAdmin": {
            "type": "boolean"
          },
          "githubUsername": {
            "type": "string",
            "nullable": true
          },
          "creationTime": {
            "type": "string",
            "format": "date-time"
          },
          "lastOnlineTime": {
            "type": "string",
            "format": "date-time"
          },
          "barsPoints": {
            "type": "integer",
            "format": "int32"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "StudySemester": {
        "type": "string",
        "enum": [
          "Y19H2",
          "Y20H1",
          "Y20H2",
          "Y21H1"
        ]
      },
      "StudentGraduationYear": {
        "type": "string",
        "enum": [
          "Undefined",
          "Y20",
          "Y21",
          "Y22",
          "Y23",
          "Y24"
        ]
      },
      "StudyProgram": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "StudyCourse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "graduationYear": {
            "$ref": "#/components/schemas/StudentGraduationYear"
          },
          "studyProgramId": {
            "type": "integer",
            "format": "int32"
          },
          "studyProgram": {
            "$ref": "#/components/schemas/StudyProgram"
          }
        }
      },
...
      "StudyGroup": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "groupName": {
            "type": "string",
            "nullable": true
          },
          "studyCourseId": {
            "type": "integer",
            "format": "int32"
          },
          "studyCourse": {
            "$ref": "#/components/schemas/StudyCourse"
          },
          "groupAdminId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "students": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Student"
            }
          },
          "groupSubjects": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/GroupSubject"
            }
          }
        }
      },
      "UniversitySystemUser": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "middleName": {
            "type": "string",
            "nullable": true
          },
          "secondName": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "GroupSubject": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "subjectId": {
            "type": "integer",
            "format": "int32"
          },
          "subject": {
            "$ref": "#/components/schemas/Subject"
          },
          "studySemester": {
            "$ref": "#/components/schemas/StudySemester"
          },
          "studyGroupId": {
            "type": "integer",
            "format": "int32"
          },
          "studyGroup": {
            "$ref": "#/components/schemas/StudyGroup"
          },
          "lectorTeacherId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "lectorTeacher": {
            "$ref": "#/components/schemas/UniversitySystemUser"
          },
          "practiceTeacherId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "practiceTeacher": {
            "$ref": "#/components/schemas/UniversitySystemUser"
          },
          "serializedGoogleTableConfig": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "Subject": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "groupSubjects": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/GroupSubject"
            }
          }
        }
      },
      "AssignmentInfoDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "creationTimeUtc": {
            "type": "string",
            "format": "date-time"
          },
          "deadlineTimeUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "creator": {
            "$ref": "#/components/schemas/IwentysUserInfoDto"
          },
          "subject": {
            "$ref": "#/components/schemas/Subject"
          },
          "isCompeted": {
            "type": "boolean"
          }
        }
      },
      "AssignmentCreateArguments": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "subjectId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "deadlineTimeUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "forStudyGroup": {
            "type": "boolean"
          },
          "link": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "CodingActivityInfoResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "month": {
            "type": "string",
            "nullable": true
          },
          "activity": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "QuestState": {
        "type": "string",
        "enum": [
          "Active",
          "Completed",
          "Revoked"
        ]
      },
      "QuestResponseInfoDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "student": {
            "$ref": "#/components/schemas/IwentysUserInfoDto"
          },
          "responseTime": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "QuestInfoDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "type": "integer",
            "format": "int32"
          },
          "creationTime": {
            "type": "string",
            "format": "date-time"
          },
          "deadline": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "state": {
            "$ref": "#/components/schemas/QuestState"
          },
          "isOutdated": {
            "type": "boolean"
          },
          "author": {
            "$ref": "#/components/schemas/IwentysUserInfoDto"
          },
          "executor": {
            "$ref": "#/components/schemas/IwentysUserInfoDto"
          },
          "responseInfos": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/QuestResponseInfoDto"
            }
          }
        }
      },
      "CreateQuestRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "type": "integer",
            "format": "int32"
          },
          "deadline": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
      "StudentUpdateRequestDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "githubUsername": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "TournamentType": {
        "type": "string",
        "enum": [
          "CodeMarathon"
        ]
      },
      "TournamentTeamInfoDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "teamName": {
            "type": "string",
            "nullable": true
          },
          "registrationTime": {
            "type": "string",
            "format": "date-time"
          },
          "points": {
            "type": "integer",
            "format": "int32"
          },
          "memberIds": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "integer",
              "format": "int32"
            }
          }
        }
      },
      "TournamentInfoResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "type": {
            "$ref": "#/components/schemas/TournamentType"
          },
          "teams": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TournamentTeamInfoDto"
            }
          }
        }
      },
...
    }
  }
}